backend/requirements.txt
crewai==0.76.0
crewai-tools==0.14.0
langchain==0.3.7
langchain-anthropic==0.2.4
langchain-openai==0.2.5
langgraph==0.2.28
langchain-community==0.3.7
chromadb==0.5.18
mem0ai==0.1.29
fastapi==0.115.4
uvicorn==0.32.0
websockets==13.1
pydantic==2.9.2
pydantic-settings==2.6.1
python-dotenv==1.0.1
celery==5.4.0
redis==5.2.0
playwright==1.48.0
openpyxl==3.1.5
apscheduler==3.10.4
twilio==9.3.6
tavily-python==0.5.0
yfinance==0.2.44
pandas==2.2.3
matplotlib==3.9.2
google-api-python-client==2.151.0
google-auth-oauthlib==1.2.1
supabase==2.9.1
httpx==0.27.2
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
email-validator==2.2.0
